@@ -1,12 +1,12 @@
 name            : example
 description     : Example Project for animated_overflow.
-version         : 1.0.2
+version         : 2.0.0
 author          : Ali Yigit Bireroglu <aliyigitbireroglu@gmail.com>
 repository      : https://github.com/aliyigitbireroglu/flutter-animated-overflow
 homepage        : https://www.cosmossoftware.coffee
 
 environment     :
-  sdk: ">=2.2.2 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies    :
   flutter          :
