@@ -1,17 +1,17 @@
 name            : animated_overflow
 description     : A simple widget for animating overflowing content horizontally or vertically with fade.
-version         : 1.0.2
+version         : 2.0.0
 author          : Ali Yigit Bireroglu <aliyigitbireroglu@gmail.com>
 repository      : https://github.com/aliyigitbireroglu/flutter-animated-overflow
 homepage        : https://www.cosmossoftware.coffee
 
 environment     :
-  sdk: ">=2.2.2 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies    :
   flutter          :
     sdk: flutter
-  simple_animations: ^2.2.1
+  simple_animations: ^3.1.1
 
 dev_dependencies:
   flutter_test:
