@@ -25,9 +25,9 @@ class MyApp extends StatelessWidget {
 }
 
 class MyHomePage extends StatefulWidget {
-  MyHomePage({Key key, this.title}) : super(key: key);
+  MyHomePage({Key? key, this.title}) : super(key: key);
 
-  final String title;
+  final String? title;
 
   @override
   _MyHomePageState createState() => _MyHomePageState();
@@ -41,7 +41,7 @@ class _MyHomePageState extends State<MyHomePage> {
       backgroundColor: const Color(0xFF29292A),
       appBar: AppBar(
         backgroundColor: const Color(0xFF1B1D1E),
-        title: Text(widget.title),
+        title: Text(widget.title!),
       ),
       body: Column(
         crossAxisAlignment: CrossAxisAlignment.start,
